plugged new racing.game template
