+++ packages/@coorpacademy-app-racing/src/web/views/index.js
@@ -1,6 +1,5 @@
 import Loading from '@coorpacademy/components/es/template/app-player/loading';
-import Question from '@coorpacademy/components/es/template/app-racing/question';
-import Race from '@coorpacademy/components/es/template/app-racing/race';
+import Game from '@coorpacademy/components/es/template/app-racing/game';
 import pipe from 'lodash/fp/pipe';
 import {createElement} from 'react';
 
@@ -9,6 +8,5 @@ export const createStateToVNode = (template, mapStateToProps) =>
 
 export const views = {
   loading: Loading,
-  question: Question,
-  race: Race
+  game: Game
 };
